Fix: 첨부파일 관련 util 분리#39
