+++ src/types/attachedFile.ts
@@ -11,10 +11,3 @@ export interface AttachedFile {
   uploadDate: string;
   description?: string;
 }
-export function getFileIcon(type: string) {
-  if (type.includes("pdf")) return "📄";
-  if (type.includes("excel") || type.includes("spreadsheet")) return "📊";
-  if (type.includes("word") || type.includes("document")) return "📝";
-  if (type.includes("image")) return "🖼️";
-  return "📎";
-}
